+++ src/dxtbx/util/dlsnxs2cbf.py
@@ -1,19 +1,12 @@
 import binascii
 
 import h5py
+import hdf5plugin  # noqa; F401
 import numpy as np
 
 from scitbx.array_family import flex
 
 from dxtbx.ext import compress
-
-try:
-    import hdf5plugin
-
-    assert hdf5plugin
-except ModuleNotFoundError:
-    # Optional dependency that can also be satisfied by hdf5-external-filter-plugins
-    pass
 
 
 def get_mask(nfast, nslow):
@@ -81,8 +74,8 @@ def compute_cbf_header(f, nn=0):
     timestamp = f["/entry/start_time"][()].decode()[:19]
     omega = f["/entry/sample/transformations/omega"][()]
     omega_increment = f["/entry/sample/transformations/omega_increment_set"][()]
-    chi = f["/entry/sample/transformations/chi"][()]
-    phi = f["/entry/sample/transformations/phi"][()]
+    chi_key = "/entry/sample/transformations/chi"
+    phi_key = "/entry/sample/transformations/phi"
 
     if "/entry/instrument/detector/beam_centre_x" in f:
         Bx = instrument["detector/beam_centre_x"][()]
@@ -124,12 +117,14 @@ _array_data.header_contents
     result.append("# Polarization 0.990")
     result.append("# Alpha 0.0000 deg.")
     result.append("# Kappa 0.0000 deg.")
-    result.append("# Phi %.4f deg." % phi)
-    result.append("# Phi_increment 0.0000 deg.")
+    if phi_key in f:
+        result.append(f"# Phi {np.squeeze(f[phi_key][()]):.4f} deg.")
+        result.append("# Phi_increment 0.0000 deg.")
     result.append("# Omega %.4f deg." % omega[nn])
     result.append("# Omega_increment %.4f deg." % omega_increment[nn])
-    result.append("# Chi %.4f deg." % chi)
-    result.append("# Chi_increment 0.0000 deg.")
+    if chi_key in f:
+        result.append(f"# Chi {np.squeeze(f[chi_key][()]):.4f} deg.")
+        result.append("# Chi_increment 0.0000 deg.")
     result.append("# Oscillation_axis X.CW")
     result.append("# N_oscillations 1")
     result.append(";")